% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape.R
\name{get_pedestrian}
\alias{get_pedestrian}
\title{API to the pedestrian data sourced from the City of Melbourne}
\usage{
get_pedestrian(from = to - 6L, to = Sys.Date() - 1L, tz = "")
}
\arguments{
\item{from}{Starting date.}

\item{to}{Ending date.}

\item{tz}{Time zone.}
}
\value{
A data frame contains "Sensor", "Date_Time", "Count".
}
\description{
API to the pedestrian data sourced from the City of Melbourne
}
\examples{
  ped_df <- get_pedestrain()
}
