% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{melb_shine}
\alias{melb_shine}
\title{A simple shiny app for pedestrian data}
\usage{
melb_shine()
}
\value{
A shiny app.
}
\description{
Provides a GUI to download data of selected sensors over a specified period
as a CSV file, accompanied with basic visualisation.
}
\details{
It offers some basic plots to give a glimpse of the data over a
short time period. In order to be reproducible, scripting using \code{walk_melb}
or \code{run_melb} is recommended.
}
\seealso{
\link{melb_walk}, \link{melb_walk_fast}
}
