% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weather.R
\name{pull_weather_sensors}
\alias{pull_weather_sensors}
\title{API using Socrata to extract Melbourne microclimate sensor locations}
\usage{
pull_weather_sensors(app_token = NULL)
}
\arguments{
\item{app_token}{Characters giving the application token. A limited number of
requests can be made without an app token (\code{NULL}), but they are subject
to much lower throttling limits than request that do include one. Sign up
for an app token \href{https://data.melbourne.vic.gov.au/profile/app_tokens}{here}.}
}
\description{
API using Socrata to extract Melbourne microclimate sensor locations
}
\details{
Extract all available climate sensor types and their identifiers
\href{https://dev.socrata.com/foundry/data.melbourne.vic.gov.au/irqv-hjr4}{Socrata}.
}
\examples{
\dontrun{
pull_weather_types()
}
}
\seealso{
\link{melb_weather}
}
