% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weather.R
\name{pull_weather_types}
\alias{pull_weather_types}
\title{API using Socrata to Melbourne microclimate measurement types}
\usage{
pull_weather_types(app_token = NULL)
}
\arguments{
\item{app_token}{Characters giving the application token. A limited number of
requests can be made without an app token (\code{NULL}), but they are subject
to much lower throttling limits than request that do include one. Sign up
for an app token \href{https://data.melbourne.vic.gov.au/profile/app_tokens}{here}.}
}
\description{
API using Socrata to Melbourne microclimate measurement types
}
\details{
Extract all available climate sensor types and their identifiers
\href{https://dev.socrata.com/foundry/data.melbourne.vic.gov.au/u4vh-84j8}{Socrata}.
}
\examples{
\dontrun{
pull_weather_types()
}
}
\seealso{
\link{melb_walk_fast}
}
