% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{shine_melb}
\alias{shine_melb}
\title{A simple shiny app for pedestrian data}
\usage{
shine_melb()
}
\value{
A shiny app.
}
\description{
A simple shiny app for pedestrian data
}
\examples{
\dontrun{
  shine_melb()
}
}
