% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape.R
\name{walk_melb}
\alias{walk_melb}
\title{API to the pedestrian data sourced from the City of Melbourne}
\usage{
walk_melb(from = to - 6L, to = Sys.Date() - 1L, tz = "")
}
\arguments{
\item{from}{Starting date.}

\item{to}{Ending date.}

\item{tz}{Time zone.}
}
\value{
A data frame including "Sensor", "Date_Time", "Date", "Time",
"Count" variables.
}
\description{
API to the pedestrian data sourced from the City of Melbourne
}
\details{
The data is sourced from \href{https://data.melbourne.vic.gov.au/Transport-Movement/Pedestrian-volume-updated-monthly-/b2ak-trbp}{Melbourne Open Data Portal}.
}
\examples{
\dontrun{
  ped_df <- walk_melb()
}
}
