% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soda.R
\name{pull_sensor}
\alias{pull_sensor}
\title{API using Socrata to Melbourne pedestrian sensor locations}
\usage{
pull_sensor(app_token = NULL)
}
\arguments{
\item{app_token}{Characters giving the application token. A limited number of
requests can be made without an app token (\code{NULL}), but they are subject
to much lower throttling limits than request that do include one. Sign up
for an app token \href{https://data.melbourne.vic.gov.au/profile/app_tokens}{here}.}
}
\description{
Provides API using Socrata to Melbourne pedestrian sensor locations.
}
\details{
It provides API using \href{https://data.melbourne.vic.gov.au/resource/h57g-5234}{Socrata}.
}
\examples{
\dontrun{
pull_sensor()
}
}
\seealso{
\link{melb_walk_fast}
}
