% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{shine_melb}
\alias{shine_melb}
\title{A simple shiny app for pedestrian data}
\usage{
shine_melb()
}
\value{
A shiny app.
}
\description{
Provides a GUI to download data of selected sensors over a specified period
as a CSV file, accompanied with basic visualisation.
}
\examples{
\dontrun{
  shine_melb()
}
}
