% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{shine_melb}
\alias{shine_melb}
\title{A simple shiny app for pedestrian data}
\usage{
shine_melb()
}
\value{
A shiny app.
}
\description{
Provides a GUI to download data of selected sensors over a specified period
as a CSV file, accompanied with basic visualisation.
}
\details{
It offers some basic plots to give a glimpse of the data over a
short time period, and so it uses \link{walk_melb} with \code{tweak = TRUE} to pull the
data. In order to be reproducible, scripting using \code{walk_melb} or \code{run_melb} is
recommended.
}
\examples{
\dontrun{
  shine_melb()
}
}
\seealso{
\link{walk_melb}, \link{run_melb}
}
