% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{shine_melb}
\alias{shine_melb}
\title{Visual summaries for pedestrian data of a month window}
\usage{
shine_melb()
}
\value{
A shiny app.
}
\description{
Visual summaries for pedestrian data of a month window
}
\examples{
\dontrun{
  shine_melb()
}
}
